+++ modules/executors/inputs/list_node.py
@@ -1,5 +1,5 @@
 from modules.common import try_solve_files
-from modules.parser import solve_templates
+from modules.formatter import solve_templates
 from modules.executors.common import solve_placeholders
 from modules.executors.common import GenericExecutor
