+++ modules/executors/inputs/node_inputs.py
@@ -1,6 +1,6 @@
 
 from modules.common import try_solve_files
-from modules.parser import solve_templates
+from modules.formatter import solve_templates
 from modules.executors.common import solve_placeholders
 from modules.executors.common import GenericExecutor
 from .list_node import ListNode
