+++ modules/executors/llm/prompt_builder_node.py
@@ -1,7 +1,7 @@
 from modules.formatter import PromptBuilder
 from modules.executors.common import GenericExecutor
 from modules.executors.common import solve_placeholders
-from modules.parser import solve_templates
+from modules.formatter import solve_templates
 
 from modules.executors.common import BaseGuiParser
