+++ modules/formatter/prompt_builder.py
@@ -75,8 +75,8 @@ class Formatter:
         elif model_name.lower().find("phi-4") >= 0:
             formatter={}
             formatter["bos"]=""
-            formatter["role_string"] = {"user":"<|im_start|>user<|im_sep|>\n","assistant":"<|im_start|>assistant<|im_sep|>\n","system":"<|im_start|>system<|im_sep|>\n"}
-            formatter["role_eom"] = {"user":"<|im_end|>\n","assistant":"<|im_end|>\n","system":"<|im_end|>\n"}
+            formatter["role_string"] = {"user":"<|im_start|>user<|im_sep|>","assistant":"<|im_start|>assistant<|im_sep|>","system":"<|im_start|>system<|im_sep|>"}
+            formatter["role_eom"] = {"user":"<|im_end|>","assistant":"<|im_end|>","system":"<|im_end|>"}
             formatter["enable_system"] = True
             formatter["roles"] = ["raw","system","user","assistant"]
