+++ modules/graph/graph_node.py
@@ -1,6 +1,6 @@
 from .executor_factory import ExecutorFactory
 from modules.executors.common import GenericExecutor
-from ..parser import solve_templates
+from ..formatter import solve_templates
 import sys
 import threading
 from .common import GraphException
